pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["Maksim Ilmast <maksim.ilmast@helsinki.fi>"]

[tool.poetry.dependencies]
python = ">=3.6,<4"
dearpygui = "^0.6.365"
invoke = "^1.5.0"
coverage = "^5.5"
pytest = "^6.2.3"
requests = "^2.25.1"
Pillow = "^8.2.0"
passlib = "^1.7.4"
python-dotenv = "^0.17.1"
py-notifier = "^0.3.0"

[tool.poetry.dev-dependencies]
pylint = "^2.7.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
